ipq806x: adds the ipq-aux-loader
This adds the ipq-aux-loader from Serhii Serhieiev <adron@mstnt.com>
It is a small auxiliary loader for legacy or LZMA compressed FIT images.
This is needed to add support for Mikrotik RouterBoards as they expect a
specific ELF image for boot. The aux-loader lets you use a LZMA compressed
FIT image and does decompression of a kernel image and then boot from it.

This version of aux-loader has support for rb3011-initializations.
These initializations are extremely important for correct work
of the DTS driver and the both switch chips. Also, these
initializations turn off the LCD screen.